+++ wemake_python_styleguide/visitors/base/visitor.py
@@ -4,25 +4,34 @@ from ast import NodeVisitor
 from typing import List
 
 from wemake_python_styleguide.errors import BaseStyleViolation
+from wemake_python_styleguide.types import ConfigurationOptions
 
 
 class BaseNodeVisitor(NodeVisitor):
-    """This class allows to store errors while traversing node tree."""
+    """
+    This class allows to store errors while traversing node tree.
+
+    Attributes:
+        options: contains the options objects passed and parsed by `flake8`.
+        errors: list of errors for the specific checker.
+
+    """
+
+    options: ConfigurationOptions
 
     def __init__(self) -> None:
         """Creates new visitor instance."""
         super().__init__()
-        self._errors: List[BaseStyleViolation] = []
-
-    @property
-    def errors(self) -> List[BaseStyleViolation]:
-        """Return errors collected by this visitor."""
-        return self._errors
+        self.errors: List[BaseStyleViolation] = []
 
     def add_error(self, error: BaseStyleViolation) -> None:
         """Adds error to the visitor."""
-        self._errors.append(error)
+        self.errors.append(error)
 
-    def provide_options(self, options) -> None:
-        """Provides options for checking."""
+    def provide_options(self, options: ConfigurationOptions) -> None:
+        """
+        Provides options for checking.
+
+        It is done separately to make testing easy.
+        """
         self.options = options